requirements.txt
Flask
Flask-SocketIO
Flask-Cors
eventlet
paho-mqtt
requests
psycopg2-binary